enabled emoji moving

+++ databaseloader.cpp
@@ -211,7 +211,7 @@ QList<DatabaseLoader::EmojiGroupInfo> DatabaseLoader::listEmojiGroups() const
 
 SymbolListModel *DatabaseLoader::createEmojiGroupModel(int groupID, QObject *modelParent) const
 {
-	SymbolListModel *updateModel = new SymbolListModel(modelParent);
+	SymbolListModel *updateModel = new SymbolListModel(modelParent, true);
 	if(this->createEmojiGroupModel(groupID, updateModel))
 		return updateModel;
 	else {
@@ -250,6 +250,50 @@ bool DatabaseLoader::removeEmoji(int groupID, uint code)
 	return query.exec();
 }
 
+bool DatabaseLoader::moveEmoji(int groupID, uint code, uint before)
+{
+	QSqlQuery listQuery(this->mainDB);
+	listQuery.prepare(QStringLiteral("SELECT EmojiMapping.EmojiID, EmojiMapping.SortHint FROM EmojiGroups INNER JOIN EmojiMapping ON EmojiGroups.ID = EmojiMapping.GroupID WHERE EmojiGroups.ID = :groupID ORDER BY EmojiMapping.SortHint ASC"));
+	listQuery.bindValue(QStringLiteral(":groupID"), groupID);
+	if(!listQuery.exec())
+		return false;
+	if(!this->mainDB.transaction())
+		return false;
+
+	QList<uint> codeList;
+	bool isCollecting = false;
+	while(listQuery.next()) {
+		uint idCode = listQuery.value(0).toUInt();
+		if(idCode == code || idCode == before)
+			isCollecting = true;
+		if(isCollecting) {
+			codeList.append(idCode);
+			QSqlQuery removeQuery(this->mainDB);
+			removeQuery.prepare(QStringLiteral("DELETE FROM EmojiMapping WHERE SortHint = :hint"));
+			removeQuery.bindValue(QStringLiteral(":hint"), listQuery.value(1).toInt());
+			if(!removeQuery.exec()) {
+				this->mainDB.rollback();
+				return false;
+			}
+		}
+	}
+
+	codeList.removeOne(code);
+	codeList.insert(codeList.indexOf(before), code);
+	for(uint insCode : codeList) {
+		QSqlQuery insertQuery(this->mainDB);
+		insertQuery.prepare(QStringLiteral("INSERT INTO EmojiMapping (GroupID, EmojiID) VALUES(:group, :emoji)"));
+		insertQuery.bindValue(QStringLiteral(":group"), groupID);
+		insertQuery.bindValue(QStringLiteral(":emoji"), insCode);
+		if(!insertQuery.exec()) {
+			this->mainDB.rollback();
+			return false;
+		}
+	}
+
+	return this->mainDB.commit();
+}
+
 int DatabaseLoader::createEmojiGroup(const QString &name)
 {
 	QSqlQuery query(this->mainDB);
@@ -263,7 +307,8 @@ int DatabaseLoader::createEmojiGroup(const QString &name)
 
 bool DatabaseLoader::deleteEmojiGroup(int groupID)
 {
-	this->mainDB.transaction();
+	if(!this->mainDB.transaction())
+		return false;
 	QSqlQuery removeMappingQuery(this->mainDB);
 	removeMappingQuery.prepare(QStringLiteral("DELETE FROM EmojiMapping WHERE GroupID = :group"));
 	removeMappingQuery.bindValue(QStringLiteral(":group"), groupID);
@@ -284,7 +329,8 @@ bool DatabaseLoader::deleteEmojiGroup(int groupID)
 
 void DatabaseLoader::updateEmojiGroupOrder(const QList<int> &idOrder)
 {
-	this->mainDB.transaction();
+	if(!this->mainDB.transaction())
+		return;
 	for(int i = 0; i < idOrder.size(); i++) {
 		QSqlQuery query(this->mainDB);
 		query.prepare(QStringLiteral("UPDATE EmojiGroups SET SortHint = :index WHERE ID = :id"));
